refactor: separate render utils from API
